Updated tooltip rendering
Menus with tooltips will now always render the box, regardless of whether there's text in it or not. If no item has a tooltip the box will not be rendered (gives option menus more room). Fully custom and generic menus must now opt-in to drawing tooltips to prevent them from breaking.